some buildfile related fixups
