--- CondCore/ESSources/bin/BuildFile.xml
@@ -1,1 +1,0 @@
-<use name="FWCore/ParameterSet"/>
