--- FastSimulation/CTPPSFastTrackingProducer/BuildFile.xml
@@ -1,13 +1,0 @@
-<use name="FWCore/Framework"/>
-<use name="FWCore/PluginManager"/>
-<use name="FWCore/ParameterSet"/>
-<use name="FWCore/ServiceRegistry"/>
-<use name="FWCore/Utilities"/>
-<use name="Utilities/PPS"/>
-<use name="FastSimulation/CTPPSFastGeometry"/>
-<use name="FastSimDataFormats/CTPPSFastSim"/>
-<use name="DataFormats/Common"/>
-<use name="clhep"/>
-<use name="hector"/>
-<use name="root"/>
-<flags EDM_PLUGIN="1"/>
