--- FastSimulation/L1CaloTriggerProducer/BuildFile.xml
@@ -1,20 +1,0 @@
-<flags EDM_PLUGIN="1"/>
-<use name="FWCore/PluginManager"/>
-<use name="FastSimDataFormats/External"/>
-<use name="FWCore/Framework"/>
-<use name="FWCore/ParameterSet"/>
-<use name="FWCore/MessageLogger"/>
-<use name="DataFormats/JetReco"/>
-<use name="DataFormats/L1Trigger"/>
-<use name="DataFormats/CaloTowers"/>
-<use name="Geometry/CaloGeometry"/>
-<use name="Geometry/CaloTopology"/>
-<use name="DataFormats/Candidate"/>
-<use name="DataFormats/EcalRecHit"/>
-<use name="DataFormats/EcalDetId"/>
-<use name="DataFormats/EcalDigi"/>
-<use name="DataFormats/HcalDigi"/>
-<use name="DataFormats/GeometryVector"/>
-<use name="CondFormats/L1TObjects"/>
-<use name="DataFormats/HcalRecHit"/>
-<use name="Geometry/Records"/>
