+++ Utilities/ReleaseScripts/scripts/git-publish
@@ -455,7 +455,7 @@ CMSSW_CATEGORIES={"Operations": ["Configuration/Applications",
   "FastSimulation/Configuration", "FastSimulation/EgammaElectronAlgos",
   "FastSimulation/Event", "FastSimulation/EventProducer",
   "FastSimulation/ForwardDetectors", "FastSimulation/HighLevelTrigger",
-  "FastSimulation/L1CaloTriggerProducer", "FastSimulation/MaterialEffects",
+  "FastSimulation/MaterialEffects",
   "FastSimulation/MuonSimHitProducer", "FastSimulation/Muons",
   "FastSimulation/Particle", "FastSimulation/ParticleDecay",
   "FastSimulation/ParticleFlow", "FastSimulation/ParticlePropagator",
